+++ compat_go20.go
@@ -5,6 +5,11 @@ package errors
 
 import (
 	"fmt"
+	"unsafe"
 )
 
 var formatString = fmt.FormatString //nolint:gochecknoglobals
+
+func slicesEqual(a []uintptr, b []uintptr) bool {
+	return len(a) == len(b) && unsafe.SliceData(a) == unsafe.SliceData(b)
+}